Change miner back to seed. Place blocknotify etc
remove marketmaker, as this is now inside the container with komodod. 
Change the miner back to the seed node, this is less likley to crash, the mining bug is well and truly fixed now.
Change TxBlaster to not restart on fail, as this won't to anything. Script triggers on block heights, if it fails it fails.